fetch post from firebase
