+++ pages/[postId]/postId.props.ts
@@ -1,0 +1,5 @@
+import { PostProps } from "../../utils/props";
+
+export interface BlogPostProps {
+    post: PostProps,
+}
